requirements.txt
aioredis==1.3.1
argon2-cffi==20.1.0
asgi-redis==1.4.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
bson==0.5.8
certifi==2020.6.20
cffi==1.14.5
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.4.7
daphne==2.4.1
dataclasses==0.6
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-staticfiles==1.2.1
django-use-email-as-username==1.1.1
djangorestframework==3.11.0
gunicorn==20.0.4
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
priority==1.3.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.10.1
pylint==2.6.0
pylint-django==2.0.15
pylint-plugin-utils==0.6
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
redis==2.10.6
requests==2.24.0
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.2.4
toml==0.10.2
twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==18.8.1
typed-ast==1.4.1
urllib3==1.25.7
wrapt==1.11.2
zope.interface==5.3.0
